refactor: updated `Navbar` and `Search` components
